+++ voxelbotutils/cogs/utils/custom_bot.py
@@ -939,11 +939,23 @@ class Bot(commands.AutoShardedBot):
     async def start(self, token: str = None, *args, **kwargs):
         """:meta private:"""
 
+        # Get the recommended shard count for this bot
+        recommended_shard_count, _ = await self.http.get_bot_gateway()
+        self.logger.info(f"Recommended shard count for this bot: {recommended_shard_count}")
+        if recommended_shard_count / 2 > self.shard_count:
+            self.logger.warning((
+                f"The shard count for this bot ({self.shard_count}) is significantly "
+                f"lower than the recommended number {recommended_shard_count}."
+            ))
+
+        # See if we should run the startup method
         if self.config.get('database', {}).get('enabled', False):
             self.logger.info("Running startup method")
             self.startup_method = self.loop.create_task(self.startup())
         else:
             self.logger.info("Not running bot startup method due to database being disabled")
+
+        # And run the original
         self.logger.info("Running original D.py start method")
         await super().start(token or self.config['token'], *args, **kwargs)